# Implement special min/max accumulator for Strings and Binary (10% faster for Clickbench Q28)