… datafusion/sqllogictest/test_files/aal.slt
… datafusion/sqllogictest/test_files/aggregate.slt