## Which issue does this PR close?
Closes https://github.com/apache/datafusion/issues/6906

## Rationale for this change

Ensure we can turn on StringView -- see https://github.com/apache/datafusion/issues/6906. Also bonus make some queries faster

## What changes are included in this PR?

1. Implement specialized GroupsAccumulator for Strings/Binary

## Are these changes tested?
Yes -- new functional tests

### Performance

####  3x faster for some pathological queries)
For this query defined in https://github.com/apache/datafusion/issues/6906#issuecomment-2355857689

```sql
set datafusion.execution.parquet.schema_force_view_types = true;

SELECT REGEXP_REPLACE("Referer", '^https?://(?:www\\.)?([^/]+)/.*$', '\\1') AS k, AVG(length("Referer")) AS l, COUNT(*) AS c, MIN("Referer")
FROM hits_partitioned
WHERE "Referer" <> '' GROUP BY k HAVING COUNT(*) > 100000 ORDER BY l DESC LIMIT 25;
```

Run via `datafusion-cli -f q28.sql`

* main: `Elapsed 18.549 seconds.`
* This branch: `Elapsed 6.187 seconds.`

####  10% faster for normal ClickBench Q28

Q28 (with `MIN("Referrer")` also gets slightly faster:
```sql
SELECT REGEXP_REPLACE("Referer", '^https?://(?:www\.)?([^/]+)/.*$', '\1') AS k, AVG(length("Referer")) AS l, COUNT(*) AS c, MIN("Referer") FROM hits WHERE "Referer" <> '' GROUP BY k HAVING COUNT(*) > 100000 ORDER BY l DESC LIMIT 25;
```

Total results:
```
--------------------
Benchmark clickbench_partitioned.json
--------------------
┏━━━━━━━━━━━━━━┳━━━━━━━━━━━━┳━━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━┓
┃ Query        ┃  main_base ┃ alamb_min_max_strings ┃        Change ┃
┡━━━━━━━━━━━━━━╇━━━━━━━━━━━━╇━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━┩
│ QQuery 0     │     2.28ms │                2.20ms │     no change │
│ QQuery 1     │    37.05ms │               38.20ms │     no change │
│ QQuery 2     │    92.34ms │               94.39ms │     no change │
│ QQuery 3     │   100.18ms │              102.79ms │     no change │
│ QQuery 4     │   922.93ms │              917.95ms │     no change │
│ QQuery 5     │  1023.47ms │              989.25ms │     no change │
│ QQuery 6     │    34.02ms │               33.54ms │     no change │
│ QQuery 7     │    41.22ms │               42.80ms │     no change │
│ QQuery 8     │  1403.10ms │             1384.97ms │     no change │
│ QQuery 9     │  1345.01ms │             1310.50ms │     no change │
│ QQuery 10    │   333.22ms │              334.38ms │     no change │
│ QQuery 11    │   388.83ms │              390.41ms │     no change │
│ QQuery 12    │  1078.98ms │             1095.45ms │     no change │
│ QQuery 13    │  1752.46ms │             1661.97ms │ +1.05x faster │
│ QQuery 14    │  1235.38ms │             1238.51ms │     no change │
│ QQuery 15    │  1073.12ms │             1094.49ms │     no change │
│ QQuery 16    │  2536.49ms │             2511.71ms │     no change │
│ QQuery 17    │  2335.03ms │             2338.92ms │     no change │
│ QQuery 18    │  5026.47ms │             5021.49ms │     no change │
│ QQuery 19    │    93.42ms │               94.42ms │     no change │
│ QQuery 20    │  1753.61ms │             1757.55ms │     no change │
│ QQuery 21    │  2022.70ms │             2021.02ms │     no change │
│ QQuery 22    │  5226.57ms │             5197.39ms │     no change │
│ QQuery 23    │ 10480.78ms │            10503.06ms │     no change │
│ QQuery 24    │   575.37ms │              585.16ms │     no change │
│ QQuery 25    │   497.50ms │              479.91ms │     no change │
│ QQuery 26    │   660.02ms │              655.23ms │     no change │
│ QQuery 27    │  2638.55ms │             2573.67ms │     no change │
│ QQuery 28    │ 15846.93ms │            14410.80ms │ +1.10x faster │
│ QQuery 29    │   516.71ms │              528.71ms │     no change │
│ QQuery 30    │  1043.51ms │             1050.48ms │     no change │
│ QQuery 31    │  1082.67ms │             1128.12ms │     no change │
│ QQuery 32    │  4224.18ms │             4305.11ms │     no change │
│ QQuery 33    │  5152.31ms │             5228.72ms │     no change │
│ QQuery 34    │  5262.44ms │             5219.20ms │     no change │
│ QQuery 35    │  1983.76ms │             1937.33ms │     no change │
│ QQuery 36    │   277.84ms │              265.44ms │     no change │
│ QQuery 37    │   119.12ms │              118.07ms │     no change │
│ QQuery 38    │   140.92ms │              139.69ms │     no change │
│ QQuery 39    │   740.00ms │              748.95ms │     no change │
│ QQuery 40    │    60.74ms │               59.34ms │     no change │
│ QQuery 41    │    50.01ms │               48.32ms │     no change │
│ QQuery 42    │    62.75ms │               63.48ms │     no change │
└──────────────┴────────────┴───────────────────────┴───────────────┘
```

## Are there any user-facing changes?
Slightly faster performance